use byte in switch statement

+++ lucene/core/src/java/org/apache/lucene/util/bkd/DocIdsWriter.java
@@ -232,7 +232,7 @@ class DocIdsWriter {
 
   /** Read {@code count} integers and feed the result directly to {@link IntersectVisitor#visit(int)}. */
   static void readInts(IndexInput in, int count, IntersectVisitor visitor) throws IOException {
-    final int bpv = in.readByte();
+    final byte bpv = in.readByte();
     switch (bpv) {
       case DELTA:
         readDeltaVInts(in, count, visitor);
